fix - file structure
- transition to a simpler architecture
