+++ .gitignore
@@ -18,6 +18,7 @@ coverage
 /cypress/screenshots/
 
 # Editor directories and files
+.env
 .vscode/*
 !.vscode/extensions.json
 .idea
