+++ package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@primeuix/themes": "^1.2.5",
+    "axios": "^1.13.1",
     "firebase": "^12.4.0",
     "pinia": "^3.0.3",
     "primeflex": "^4.0.0",
