+++ src/main.ts
@@ -1,0 +1,31 @@
+import { createApp } from 'vue';
+import { createPinia } from 'pinia';
+
+import './plugins/prime-vue/primeflex.css';
+import './assets/styles/index.css';
+import 'primeicons/primeicons.css';
+
+import { initializeApp } from 'firebase/app';
+
+import App from './App.vue';
+import router from './router';
+import { PrimeVue, primeVueConfig } from './plugins/prime-vue';
+
+const app = createApp(App);
+
+app.use(createPinia());
+app.use(router);
+app.use(PrimeVue, primeVueConfig);
+
+const firebaseConfig = {
+  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
+  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
+  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
+  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_FIREBASE_APP_ID,
+};
+
+initializeApp(firebaseConfig);
+
+app.mount('#app');
