+++ src/pages/LoginPage.vue
@@ -1,3 +1,20 @@
+<script setup lang="ts">
+import { useAuthFlow } from '@/stores/auth.flow';
+import PrimeInput from '@/components/ui/PrimeInput.vue';
+import { Button } from 'primevue';
+import { ref } from 'vue';
+import { RouterLink } from 'vue-router';
+
+const password = ref<string>('');
+const email = ref<string>('');
+
+const { handleLogin, isLoading, error } = useAuthFlow();
+
+const handleSubmit = async () => {
+  await handleLogin(email.value, password.value);
+};
+</script>
+
 <template>
   <div class="min-h-screen flex flex-column align-items-center justify-content-center">
     <div v-if="isLoading"><h1 class="text-center">Загрузка</h1></div>
@@ -22,20 +39,3 @@
     </div>
   </div>
 </template>
-
-<script setup lang="ts">
-import { useAuthFlow } from '@/features/auth/lib/auth.flow';
-import PrimeInput from '@/shared/ui/PrimeInput.vue';
-import { Button } from 'primevue';
-import { ref, watch } from 'vue';
-import { RouterLink } from 'vue-router';
-
-const password = ref<string>('');
-const email = ref<string>('');
-
-const { handleLogin, isLoading, error, clearError } = useAuthFlow();
-
-const handleSubmit = async () => {
-  await handleLogin(email.value, password.value);
-};
-</script>
