+++ src/plugins/prime-vue/config.ts
