+++ src/plugins/prime-vue/index.ts
