+++ src/stores/search.store.ts
@@ -1,0 +1,10 @@
+import { defineStore } from 'pinia';
+import { ref } from 'vue';
+
+export const useSearchStore = defineStore('search', () => {
+  const searchedMovies = ref<[]>([]);
+
+  return {
+    searchedMovies,
+  };
+});
