--- src/widgets/header/index.ts
@@ -1,1 +1,0 @@
-export { default as PrimeHeader } from './ui/PrimeHeader.vue';
